Validation on first page of createUser fixed
